requirements.txt
# Medical AI Assistant - Complete Dependencies
# All packages required for the Medical AI Assistant project

# Core Web Framework
streamlit
streamlit-chat

# LLM Providers
openai
groq

# Database
psycopg2-binary
SQLAlchemy

# Document Processing
PyPDF2
python-docx

# Machine Learning & AI
sentence-transformers
faiss-cpu
scikit-learn
torch
transformers

# Data Processing
pandas
numpy

# Text Processing
regex
nltk

# Web & HTTP
requests
urllib3
httpx

# Utilities
python-dotenv
pydantic
typing-extensions

# Caching & Storage
joblib

# Logging
loguru

# Development & Testing
pytest
pytest-cov

# Deployment
gunicorn

# Additional Dependencies
markdown
